ADD write_to_FIFO

+++ Inc/global_variables.h
@@ -9,7 +9,7 @@
 // extern void _Error_Handler(char *, int);
  extern uint8_t FLAG_interrupt_INT3;
  extern uint8_t FLAG_interrupt_PULSE; 
- extern uint8_t loop_counter; 
+ extern int16_t loop_counter; 
  extern uint8_t counter_ADC_data_ready;
 
 #ifdef __cplusplus
